feat: day03 done
this was jumbled together code, started late and
wanted to finish but did not have the mental
energy to make this efficient or good. my part1
was *not* prepared for part2 at all, my part1
decided to scan for numbers which seemed easier
while my part2 had to start from scratch to scan
for symbols.